refactor(tui): extract command filtering and help utils
